Moved test_result outside of ml_worker module
