+++ giskard/ml_worker/websocket/utils.py
@@ -15,7 +15,7 @@ from giskard.ml_worker import websocket
 from giskard.registry.registry import tests_registry
 from giskard.registry.slicing_function import SlicingFunction
 from giskard.registry.transformation_function import TransformationFunction
-from giskard.ml_worker.testing.test_result import TestMessageLevel, TestResult
+from giskard.core.test_result import TestMessageLevel, TestResult
 from giskard.ml_worker.websocket import (
     CreateSubDatasetParam,
     DatasetProcessingParam,
