+++ giskard/registry/giskard_test.py
@@ -13,7 +13,7 @@ from giskard.core.savable import Artifact
 from giskard.core.validation import configured_validate_arguments
 from giskard.exceptions.giskard_exception import python_env_exception_helper
 from giskard.registry.registry import get_object_uuid, tests_registry
-from giskard.ml_worker.testing.test_result import TestResult
+from giskard.core.test_result import TestResult
 from giskard.utils.analytics_collector import analytics
 
 DATA_PKL = "data.pkl"
