+++ giskard/testing/tests/calibration.py
@@ -6,7 +6,7 @@ import pandas as pd
 from ...datasets.base import Dataset
 from ...registry.decorators import test
 from ...registry.slicing_function import SlicingFunction
-from ...ml_worker.testing.test_result import TestResult
+from ...core.test_result import TestResult
 from ...models.base import BaseModel
 from . import debug_description_prefix
