+++ giskard/testing/tests/drift.py
@@ -14,7 +14,7 @@ from typing import List, Optional
 from giskard.datasets.base import Dataset
 from giskard.registry.decorators import test
 from giskard.registry.slicing_function import SlicingFunction
-from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
+from giskard.core.test_result import TestMessage, TestMessageLevel, TestResult
 from giskard.models.base import BaseModel
 from giskard.testing.utils.utils import check_slice_not_empty, validate_classification_label
 from . import debug_description_prefix
