+++ giskard/testing/tests/llm/ground_truth.py
@@ -4,7 +4,7 @@ from .. import debug_description_prefix
 from ....datasets.base import Dataset
 from ....llm import LLMImportError
 from ....registry.decorators import test
-from ....ml_worker.testing.test_result import TestResult
+from ....core.test_result import TestResult
 from ....models.base import BaseModel
