+++ tests/communications/test_websocket_actor_tests.py
@@ -6,7 +6,7 @@ import pytest
 from giskard import test
 from giskard.datasets.base import Dataset
 from giskard.ml_worker import websocket
-from giskard.ml_worker.testing.test_result import TestResult as GiskardTestResult, TestMessage, TestMessageLevel
+from giskard.core.test_result import TestResult as GiskardTestResult, TestMessage, TestMessageLevel
 from giskard.ml_worker.websocket import listener
 from giskard.testing.tests import debug_prefix
 from tests import utils
