+++ tests/test_programmatic.py
@@ -4,7 +4,7 @@ from giskard import test, scan
 from giskard.core.suite import Suite, SuiteInput
 from giskard.datasets.base import Dataset
 from giskard.registry.slicing_function import SlicingFunction, slicing_function
-from giskard.ml_worker.testing.test_result import TestMessageLevel
+from giskard.core.test_result import TestMessageLevel
 from giskard.models.base import BaseModel
 from giskard.testing.tests.performance import test_auc, test_f1, test_diff_f1
